https://myddleware.atlassian.net/browse/PROJ-647 export csv to specific file
